# docs/swagger.yaml
definitions:
...
  github_com_imperatorofdwelling_Full-backend_internal_domain_models_stays.Stay:
    properties:
      created_at:
        type: string
      entrance:
        type: string
      floor:
        type: string
      guests:
        type: integer
      house:
        type: string
      id:
        type: string
      is_smoking_prohibited:
        type: boolean
      location_id:
        type: string
      name:
        type: string
      number_of_bathrooms:
        type: integer
      number_of_bedrooms:
        type: integer
      number_of_beds:
        type: integer
      price:
        type: number
      rating:
        type: number
      room:
        type: string
      square:
        type: number
      street:
        type: string
      type:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  github_com_imperatorofdwelling_Full-backend_internal_domain_models_stays.StayEntity:
    properties:
      entrance:
        type: string
      floor:
        type: string
      guests:
        type: integer
      house:
        type: string
      is_smoking_prohibited:
        default: false
        type: boolean
      location_id:
        type: string
      name:
        type: string
      number_of_bathrooms:
        type: integer
      number_of_bedrooms:
        type: integer
      number_of_beds:
        type: integer
      price:
        type: number
      room:
        type: string
      square:
        type: number
      street:
        type: string
      type:
        type: string
      user_id:
        type: string
    required:
    - guests
    - house
    - location_id
    - name
    - number_of_bathrooms
    - number_of_bedrooms
    - number_of_beds
    - price
    - square
    - street
    - type
    - user_id
    type: object
  github_com_imperatorofdwelling_Full-backend_internal_domain_models_stays.StayImage:
    properties:
      created_at:
        type: string
      id:
        type: string
      image_name:
        type: string
      is_main:
        type: boolean
      stay_id:
        type: string
      updated_at:
        type: string
    type: object
...
  github_com_imperatorofdwelling_Full-backend_internal_domain_models_user.User:
    description: User details
    properties:
      avatar:
        items:
          type: integer
        type: array
      birth_date:
        $ref: '#/definitions/sql.NullTime'
      city:
        type: string
      country:
        type: string
      createdAt:
        type: string
      email:
        type: string
      gender:
        type: string
      id:
        type: string
      name:
        type: string
      national:
        type: string
      phone:
        type: string
      updatedAt:
        type: string
    type: object
...
  github_com_imperatorofdwelling_Full-backend_internal_utils_response.ResponseError:
    properties:
      error:
        type: string
    type: object
  sql.NullTime:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
paths:
...
  /stays:
    get:
      consumes:
      - application/json
      description: Get all stays
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            items:
              $ref: '#/definitions/github_com_imperatorofdwelling_Full-backend_internal_domain_models_stays.Stay'
            type: array
        "400":
          description: Error
          schema:
            $ref: '#/definitions/github_com_imperatorofdwelling_Full-backend_internal_utils_response.ResponseError'
        default:
          description: Error
          schema:
            $ref: '#/definitions/github_com_imperatorofdwelling_Full-backend_internal_utils_response.ResponseError'
      summary: Get all stays
      tags:
      - stays
...
  /stays/{stayId}:
    delete:
      consumes:
      - application/json
      description: Delete stay by id
      parameters:
      - description: stay id
        in: path
        name: stayId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: no content
          schema:
            type: string
        "400":
          description: Error
          schema:
            $ref: '#/definitions/github_com_imperatorofdwelling_Full-backend_internal_utils_response.ResponseError'
        default:
          description: Error
          schema:
            $ref: '#/definitions/github_com_imperatorofdwelling_Full-backend_internal_utils_response.ResponseError'
      summary: Delete Stay
      tags:
      - stays
    get:
      consumes:
      - application/json
      description: get stay
      parameters:
      - description: stay id
        in: path
        name: stayId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/github_com_imperatorofdwelling_Full-backend_internal_domain_models_stays.Stay'
        "400":
          description: Error
          schema:
            $ref: '#/definitions/github_com_imperatorofdwelling_Full-backend_internal_utils_response.ResponseError'
        default:
          description: Error
          schema:
            $ref: '#/definitions/github_com_imperatorofdwelling_Full-backend_internal_utils_response.ResponseError'
      summary: Get Stay by id
      tags:
      - stays
    put:
      consumes:
      - application/json
      description: Update stay by id
      parameters:
      - description: request stay data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_imperatorofdwelling_Full-backend_internal_domain_models_stays.StayEntity'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/github_com_imperatorofdwelling_Full-backend_internal_domain_models_stays.Stay'
        "400":
          description: Error
          schema:
            $ref: '#/definitions/github_com_imperatorofdwelling_Full-backend_internal_utils_response.ResponseError'
        default:
          description: Error
          schema:
            $ref: '#/definitions/github_com_imperatorofdwelling_Full-backend_internal_utils_response.ResponseError'
      summary: Update Stay
      tags:
      - stays
  /stays/create:
    post:
      consumes:
      - application/json
      description: Create stay
      parameters:
      - description: request stay data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_imperatorofdwelling_Full-backend_internal_domain_models_stays.StayEntity'
      produces:
      - application/json
      responses:
        "201":
          description: created
          schema:
            type: string
        "400":
          description: Error
          schema:
            $ref: '#/definitions/github_com_imperatorofdwelling_Full-backend_internal_utils_response.ResponseError'
        default:
          description: Error
          schema:
            $ref: '#/definitions/github_com_imperatorofdwelling_Full-backend_internal_utils_response.ResponseError'
      summary: Create Stay
      tags:
      - stays
  /stays/images:
    post:
      consumes:
      - multipart/form-data
      description: 'Create images (IMPORTANT: not main image). Form data with two
        rows: stay_id, images in array'
      parameters:
      - collectionFormat: csv
        description: images list in form data
        in: formData
        items:
          type: file
        name: images
        required: true
        type: array
      - description: stay id
        in: formData
        name: stay_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
        "400":
          description: Error
          schema:
            $ref: '#/definitions/github_com_imperatorofdwelling_Full-backend_internal_utils_response.ResponseError'
        default:
          description: Error
          schema:
            $ref: '#/definitions/github_com_imperatorofdwelling_Full-backend_internal_utils_response.ResponseError'
      summary: Create images
      tags:
      - stays
  /stays/images/{stayId}:
    get:
      consumes:
      - application/json
      description: Get stays images by stay id
      parameters:
      - description: stay id
        in: path
        name: stayId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            items:
              $ref: '#/definitions/github_com_imperatorofdwelling_Full-backend_internal_domain_models_stays.StayImage'
            type: array
        "400":
          description: Error
          schema:
            $ref: '#/definitions/github_com_imperatorofdwelling_Full-backend_internal_utils_response.ResponseError'
        default:
          description: Error
          schema:
            $ref: '#/definitions/github_com_imperatorofdwelling_Full-backend_internal_utils_response.ResponseError'
      summary: Get all stays images by stay id
      tags:
      - stays
  /stays/images/delete/{imageId}:
    delete:
      consumes:
      - application/json
      description: Delete image by id
      parameters:
      - description: stay image id
        in: path
        name: imageId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
        "400":
          description: Error
          schema:
            $ref: '#/definitions/github_com_imperatorofdwelling_Full-backend_internal_utils_response.ResponseError'
        default:
          description: Error
          schema:
            $ref: '#/definitions/github_com_imperatorofdwelling_Full-backend_internal_utils_response.ResponseError'
      summary: Delete image by id
      tags:
      - stays
  /stays/images/main:
    post:
      consumes:
      - multipart/form-data
      description: 'Create main image (IMPORTANT: should be one main image). If main
        image already exists in stay, it will be replaced with this new image. The
        old replaced image will no longer be the main and "is_main" value replaced
        to false. Request data should be in Form data with two values: images and
        stay_id. Note, that images row is one image (not array)'
      parameters:
      - description: images
        in: formData
        name: images
        required: true
        type: file
      - description: stay id
        in: formData
        name: stay_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
        "400":
          description: Error
          schema:
            $ref: '#/definitions/github_com_imperatorofdwelling_Full-backend_internal_utils_response.ResponseError'
        default:
          description: Error
          schema:
            $ref: '#/definitions/github_com_imperatorofdwelling_Full-backend_internal_utils_response.ResponseError'
      summary: Create main image
      tags:
      - stays
  /stays/images/main/{stayId}:
    get:
      consumes:
      - application/json
      description: Get main stays image by stay id
      parameters:
      - description: stay id
        in: path
        name: stayId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/github_com_imperatorofdwelling_Full-backend_internal_domain_models_stays.StayImage'
        "400":
          description: Error
          schema:
            $ref: '#/definitions/github_com_imperatorofdwelling_Full-backend_internal_utils_response.ResponseError'
        default:
          description: Error
          schema:
            $ref: '#/definitions/github_com_imperatorofdwelling_Full-backend_internal_utils_response.ResponseError'
      summary: Get main stays image by stay id
      tags:
      - stays
  /stays/location/{locationId}:
    get:
      consumes:
      - application/json
      description: get stays by location id. Handler additionally checks is location
        exist.
      parameters:
      - description: stay id
        in: path
        name: locationId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            items:
              $ref: '#/definitions/github_com_imperatorofdwelling_Full-backend_internal_domain_models_stays.Stay'
            type: array
        "400":
          description: Error
          schema:
            $ref: '#/definitions/github_com_imperatorofdwelling_Full-backend_internal_utils_response.ResponseError'
        default:
          description: Error
          schema:
            $ref: '#/definitions/github_com_imperatorofdwelling_Full-backend_internal_utils_response.ResponseError'
      summary: Get Stays by location id
      tags:
      - stays
  /stays/user/{userId}:
    get:
      consumes:
      - application/json
      description: Get stays by user id
      parameters:
      - description: user id
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            items:
              $ref: '#/definitions/github_com_imperatorofdwelling_Full-backend_internal_domain_models_stays.Stay'
            type: array
        "400":
          description: Error
          schema:
            $ref: '#/definitions/github_com_imperatorofdwelling_Full-backend_internal_utils_response.ResponseError'
        default:
          description: Error
          schema:
            $ref: '#/definitions/github_com_imperatorofdwelling_Full-backend_internal_utils_response.ResponseError'
      summary: Get all stays by user id
      tags:
      - stays
...
  /user/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a user with the provided ID
      operationId: deleteUserByID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: User successfully deleted
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/github_com_imperatorofdwelling_Full-backend_internal_utils_response.ResponseError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_imperatorofdwelling_Full-backend_internal_utils_response.ResponseError'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/github_com_imperatorofdwelling_Full-backend_internal_utils_response.ResponseError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_imperatorofdwelling_Full-backend_internal_utils_response.ResponseError'
      security:
      - ApiKeyAuth: []
      summary: Delete a user by ID
      tags:
      - users
    get:
      consumes:
      - application/json
      description: Retrieves a user by the provided ID
      operationId: getUserByID
      parameters:
      - description: User  ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_imperatorofdwelling_Full-backend_internal_domain_models_user.User'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/github_com_imperatorofdwelling_Full-backend_internal_utils_response.ResponseError'
        "404":
          description: User  not found
          schema:
            $ref: '#/definitions/github_com_imperatorofdwelling_Full-backend_internal_utils_response.ResponseError'
      summary: Get a user by ID
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Update a user with the provided ID
      operationId: updateUserByID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: User update data
        in: body
        name: model.User
        required: true
        schema:
          $ref: '#/definitions/github_com_imperatorofdwelling_Full-backend_internal_domain_models_user.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_imperatorofdwelling_Full-backend_internal_domain_models_user.User'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/github_com_imperatorofdwelling_Full-backend_internal_utils_response.ResponseError'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/github_com_imperatorofdwelling_Full-backend_internal_utils_response.ResponseError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_imperatorofdwelling_Full-backend_internal_utils_response.ResponseError'
      security:
      - ApiKeyAuth: []
      summary: Update a user by ID
      tags:
      - users
...
swagger: "2.0"